update translations, add pull-down refresh, separate top bar components to files
