+++ res/values/strings_ar.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "  (Ethereum، Polygon) ﻯﺮﺧﺃ ﺓﺮﻣ ﺔﻟﻭﺎﺤﻤﻟﺍﻭ EVM ﻊﻣ ﺔﻘﻓﺍﻮﺘﻣ ﺔﻈﻔﺤﻣ ﻰﻟﺇ ﻞﻳﺪﺒﺘﻟﺍ ﻰﺟﺮﻳ",
   "symbol": "ﺰﻣﺭ",
   "sync_all_wallets": "مزامنة جميع المحافظ",
+  "sync_attempting": "محاولة المزامنة",
+  "sync_attempting_scan": "محاولة المسح",
+  "sync_connected": "متصل",
+  "sync_connecting": "الاتصال",
+  "sync_not_connected": "غير متصل",
+  "sync_offline": "غير متصل",
+  "sync_starting_scan": "بدء المسح",
   "sync_status_attempting_scan": "محاولة المسح",
   "sync_status_attempting_sync": "جاري محاولة المزامنة",
   "sync_status_connected": "متصل",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "يتم المزامنة",
   "sync_status_timed_out": "نفد وقته",
   "sync_status_unsupported": "عقدة غير مدعومة",
+  "sync_syncing": "المزامنة",
+  "sync_timed_out": "انتهت مهلة المزامنة",
+  "sync_unsupported_node": "عقدة غير مدعومة",
   "synchronizing": "المزامنة",
   "syncing_wallet_alert_content": "قد لا يكتمل رصيدك وقائمة المعاملات الخاصة بك حتى تظهر عبارة “SYNCHRONIZED“ في الأعلى. انقر / اضغط لمعرفة المزيد.",
   "syncing_wallet_alert_title": "محفظتك تتم مزامنتها",
