+++ res/values/strings_bg.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Моля, превключете към портфейл, съвместим с EVM, и опитайте отново (Ethereum, Polygon)",
   "symbol": "Символ",
   "sync_all_wallets": "Синхронизирайте всички портфейли",
+  "sync_attempting": "Опит за синхронизиране",
+  "sync_attempting_scan": "Опит за сканиране",
+  "sync_connected": "Свързан",
+  "sync_connecting": "Свързване",
+  "sync_not_connected": "Не е свързан",
+  "sync_offline": "Офлайн",
+  "sync_starting_scan": "Стартира сканиране",
   "sync_status_attempting_scan": "Опит за сканиране",
   "sync_status_attempting_sync": "ОПИТ ЗА СИНХРОНИЗАЦИЯ",
   "sync_status_connected": "СВЪРЗВАНЕ",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "СИНХРОНИЗИРАНЕ",
   "sync_status_timed_out": "ВРЕМЕТО ИЗТЕЧЕ",
   "sync_status_unsupported": "Неподдържан възел",
+  "sync_syncing": "Синхронизиране",
+  "sync_timed_out": "Времето за изчакване на синхронизирането изтече",
+  "sync_unsupported_node": "Неподдържан възел",
   "synchronizing": "Синхронизиране",
   "syncing_wallet_alert_content": "Списъкът ви с баланс и транзакции може да не е пълен, докато в горната част не пише „СИНХРОНИЗИРАН“. Кликнете/докоснете, за да научите повече.",
   "syncing_wallet_alert_title": "Вашият портфейл се синхронизира",
