+++ res/values/strings_cs.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Přepněte na peněženku kompatibilní s EVM a zkuste to znovu (Ethereum, Polygon)",
   "symbol": "Symbol",
   "sync_all_wallets": "Synchronizovat všechny peněženky",
+  "sync_attempting": "Pokus o synchronizaci",
+  "sync_attempting_scan": "Pokus o skenování",
+  "sync_connected": "Připojeno",
+  "sync_connecting": "Spojovací",
+  "sync_not_connected": "Nepřipojeno",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Spouštění skenování",
   "sync_status_attempting_scan": "Pokus o skenování",
   "sync_status_attempting_sync": "ZAHAJUJI SYNCHR.",
   "sync_status_connected": "PŘIPOJENO",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "SYNCHRONIZUJI",
   "sync_status_timed_out": "ČAS VYPRŠEL",
   "sync_status_unsupported": "Nepodporovaný uzel",
+  "sync_syncing": "Probíhá synchronizace",
+  "sync_timed_out": "Vypršel časový limit synchronizace",
+  "sync_unsupported_node": "Nepodporovaný uzel",
   "synchronizing": "Synchronizace",
   "syncing_wallet_alert_content": "Váš seznam zůstatků a transakcí nemusí být úplný, dokud nebude nahoře uvedeno „SYNCHRONIZOVANÉ“. Kliknutím/klepnutím se dozvíte více.",
   "syncing_wallet_alert_title": "Vaše peněženka se synchronizuje",
