+++ res/values/strings_de.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Bitte wechseln Sie zu einem EVM-kompatiblen Wallet und versuchen Sie es erneut (Ethereum, Polygon)",
   "symbol": "Symbol",
   "sync_all_wallets": "Alle Wallets synchronisieren",
+  "sync_attempting": "Synchronisierungsversuch",
+  "sync_attempting_scan": "Scanversuch wird durchgeführt",
+  "sync_connected": "Verbunden",
+  "sync_connecting": "Verbinden",
+  "sync_not_connected": "Nicht verbunden",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Scan wird gestartet",
   "sync_status_attempting_scan": "Versuch Scan",
   "sync_status_attempting_sync": "SYNC VERSUCHEN",
   "sync_status_connected": "VERBUNDEN",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "SYNCHRONISIERE",
   "sync_status_timed_out": "Zeitlich abgestimmt",
   "sync_status_unsupported": "Nicht unterstützter Node",
+  "sync_syncing": "Synchronisierung",
+  "sync_timed_out": "Zeitüberschreitung bei der Synchronisierung",
+  "sync_unsupported_node": "Nicht unterstützter Knoten",
   "synchronizing": "Synchronisierung",
   "syncing_wallet_alert_content": "Ihr Kontostand und Ihre Transaktionsliste sind möglicherweise erst vollständig, wenn oben „SYNCHRONISIERT“ steht. Klicken/tippen Sie, um mehr zu erfahren.",
   "syncing_wallet_alert_title": "Ihr Wallet wird synchronisiert",
