+++ res/values/strings_en.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Please switch to an EVM compatible wallet and try again (Ethereum, Polygon)",
   "symbol": "Symbol",
   "sync_all_wallets": "Sync all wallets",
+  "sync_attempting": "Attempting sync",
+  "sync_attempting_scan": "Attempting scan",
+  "sync_connected": "Connected",
+  "sync_connecting": "Connecting",
+  "sync_not_connected": "Not Connected",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Starting scan",
   "sync_status_attempting_scan": "ATTEMPTING SCAN",
   "sync_status_attempting_sync": "ATTEMPTING SYNC",
   "sync_status_connected": "CONNECTED",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "SYNCHRONIZING",
   "sync_status_timed_out": "TIMED OUT",
   "sync_status_unsupported": "UNSUPPORTED NODE",
+  "sync_syncing": "Syncing",
+  "sync_timed_out": "Sync Timed Out",
+  "sync_unsupported_node": "Unsupported node",
   "synchronizing": "Synchronizing",
   "syncing_wallet_alert_content": "Your balance and transaction list may not be complete until it says “SYNCHRONIZED” at the top. Click/tap to learn more.",
   "syncing_wallet_alert_title": "Your wallet is syncing",
