+++ res/values/strings_es.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Cambia a una billetera compatible con EVM e inténtalo de nuevo (Ethereum, Polygon)",
   "symbol": "Símbolo",
   "sync_all_wallets": "Sincronizar todas las carteras",
+  "sync_attempting": "Intentando sincronizar",
+  "sync_attempting_scan": "Intentando escanear",
+  "sync_connected": "Conectado",
+  "sync_connecting": "Conectando",
+  "sync_not_connected": "No conectado",
+  "sync_offline": "Desconectado",
+  "sync_starting_scan": "Iniciando escaneo",
   "sync_status_attempting_scan": "INTENTANDO ESCANEAR",
   "sync_status_attempting_sync": "INTENTANDO SINCRONIZAR",
   "sync_status_connected": "CONECTADO",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "SINCRONIZANDO",
   "sync_status_timed_out": "TIEMPO AGOTADO",
   "sync_status_unsupported": "NODO NO COMPATIBLE",
+  "sync_syncing": "Sincronización",
+  "sync_timed_out": "Sincronización agotada",
+  "sync_unsupported_node": "Nodo no soportado",
   "synchronizing": "Sincronizando",
   "syncing_wallet_alert_content": "Es posible que tu saldo y lista de transacciones no estén completos hasta que aparezca “SINCRONIZADO” en la parte superior. Haga clic o pulsa para obtener más información.",
   "syncing_wallet_alert_title": "Tu cartera se está sincronizando",
