+++ res/values/strings_fa.arb
@@ -941,6 +941,13 @@
   "switchToEVMCompatibleWallet": "لطفاً به یک کیف پول سازگار با EVM تغییر دهید و دوباره تلاش کنید (اتریوم، پالیگان)",
   "symbol": "نماد",
   "sync_all_wallets": "همگام‌سازی همه کیف پول‌ها",
+  "sync_attempting": "تلاش برای همگام سازی",
+  "sync_attempting_scan": "تلاش برای اسکن",
+  "sync_connected": "متصل شد",
+  "sync_connecting": "اتصال",
+  "sync_not_connected": "متصل نیست",
+  "sync_offline": "آفلاین",
+  "sync_starting_scan": "شروع اسکن",
   "sync_status_attempting_scan": "تلاش برای اسکن",
   "sync_status_attempting_sync": "تلاش برای همگام‌سازی",
   "sync_status_connected": "متصل",
@@ -956,6 +963,9 @@
   "sync_status_syncronizing": "در حال همگام‌سازی",
   "sync_status_timed_out": "پایان مهلت زمانی",
   "sync_status_unsupported": "نود پشتیبانی نشده",
+  "sync_syncing": "در حال همگام سازی",
+  "sync_timed_out": "زمان همگام سازی تمام شد",
+  "sync_unsupported_node": "گره پشتیبانی نشده",
   "synchronizing": "در حال همگام‌سازی",
   "syncing_wallet_alert_content": "موجودی و لیست تراکنش‌های شما ممکن است کامل نباشد تا زمانی که در بالا بنویسد «همگام‌سازی شد».\nبرای اطلاعات بیشتر کلیک/لمس کنید.",
   "syncing_wallet_alert_title": "کیف پول شما در حال همگام‌سازی است",
