+++ res/values/strings_fr.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Veuillez passer à un portefeuille compatible EVM et réessayer (Ethereum, Polygon)",
   "symbol": "Symbole",
   "sync_all_wallets": "Synchroniser tous les portefeuilles",
+  "sync_attempting": "Tentative de synchronisation",
+  "sync_attempting_scan": "Tentative d'analyse",
+  "sync_connected": "Connecté",
+  "sync_connecting": "De liaison",
+  "sync_not_connected": "Non connecté",
+  "sync_offline": "Hors ligne",
+  "sync_starting_scan": "Démarrage de l'analyse",
   "sync_status_attempting_scan": "Tentative de numérisation",
   "sync_status_attempting_sync": "TENTATIVE DE SYNCHRONISATION",
   "sync_status_connected": "CONNECTÉ",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "SYNCHRONISATION EN COURS",
   "sync_status_timed_out": "FIN DU TEMPS",
   "sync_status_unsupported": "Nœud non pris en charge",
+  "sync_syncing": "Synchronisation",
+  "sync_timed_out": "La synchronisation a expiré",
+  "sync_unsupported_node": "Nœud non pris en charge",
   "synchronizing": "Synchronisation",
   "syncing_wallet_alert_content": "Votre solde et votre liste de transactions peuvent ne pas être à jour tant que la mention « SYNCHRONISÉ » n'apparaît en haut de l'écran. Cliquez/appuyez pour en savoir plus.",
   "syncing_wallet_alert_title": "Votre portefeuille (wallet) est en cours de synchronisation",
