+++ res/values/strings_gn.arb
@@ -720,6 +720,13 @@
   "switchToEVMCompatibleWallet": "Emoambue peteî billetera ojoajukuáava EVM rehe ha reñehaʼã pyahu jey (Ethereum, Polygon).",
   "symbol": "Ta'anga’i",
   "sync_all_wallets": "Ñemobojoaju enteroite billetera",
+  "sync_attempting": "Oñeha’ãvo sincronización .",
+  "sync_attempting_scan": "Oñeha'ãvo escaneo .",
+  "sync_connected": "Ojoajúva .",
+  "sync_connecting": "Ojoajúvo .",
+  "sync_not_connected": "Ndojejoajúi .",
+  "sync_offline": "Offline rehegua .",
+  "sync_starting_scan": "Oñepyrũvo escaneo .",
   "sync_status_attempting_sync": "Oñeha’ã ombojoaju",
   "sync_status_connected": "Oñembojoaju",
   "sync_status_connecting": "Oñembojoajuhína",
@@ -731,6 +738,9 @@
   "sync_status_syncronizing": "Oñembojoajuhína",
   "sync_status_timed_out": "Ndoikoveimai",
   "sync_status_unsupported": "Nodo ndaha’éi ojuehegua",
+  "sync_syncing": "Ñe’ẽjoajuha .",
+  "sync_timed_out": "Sincronización tiempo rehegua .",
+  "sync_unsupported_node": "Nodo ndojeguerohorýiva .",
   "syncing_wallet_alert_content": "Ikatú nde saldo térã ne transacción noñemohu’ãi he’í peve “joajuma” yvate gotyo. Ehecha, ejopy térã epoko reikuaavé hag̃ua.",
   "syncing_wallet_alert_title": "Ojoajuma nde billetera",
   "template": "Plantilla-kuéra",
