+++ res/values/strings_ha.arb
@@ -944,6 +944,13 @@
   "switchToEVMCompatibleWallet": "Da fatan za a canza zuwa walat ɗin EVM mai jituwa kuma a sake gwadawa (Ethereum, Polygon)",
   "symbol": "Alama",
   "sync_all_wallets": "Sync Duk Wallet",
+  "sync_attempting": "Ƙoƙarin daidaitawa",
+  "sync_attempting_scan": "Ƙoƙarin dubawa",
+  "sync_connected": "An haɗa",
+  "sync_connecting": "Haɗawa",
+  "sync_not_connected": "Ba a Haɗe ba",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Fara dubawa",
   "sync_status_attempting_scan": "Yunƙurin scan",
   "sync_status_attempting_sync": "KWAFI",
   "sync_status_connected": "HANNU",
@@ -959,6 +966,9 @@
   "sync_status_syncronizing": "KWAFI",
   "sync_status_timed_out": "ATED Out",
   "sync_status_unsupported": "Ba a Taimako ba",
+  "sync_syncing": "Ana daidaitawa",
+  "sync_timed_out": "Lokacin Daidaitawa ya ƙare",
+  "sync_unsupported_node": "Kumburi mara tallafi",
   "synchronizing": "Aikiɓaɓaƙe",
   "syncing_wallet_alert_content": "Ma'aunin ku da lissafin ma'amala bazai cika ba har sai an ce \"SYNCHRONIZED\" a saman. Danna/matsa don ƙarin koyo.",
   "syncing_wallet_alert_title": "Walat ɗin ku yana aiki tare",
