+++ res/values/strings_hi.arb
@@ -623,8 +623,8 @@
   "payjoin_unavailable_sheet_title": "Payjoin अनुपलब्ध क्यों है?",
   "payment_id": "भुगतान ID: ",
   "payment_made_easy": "भुगतान आसान किया गया",
-  "Payment_was_received": "आपका भुगतान प्राप्त हो गया था।",
   "payment_was_received": "आपका भुगतान प्राप्त हुआ था।",
+  "Payment_was_received": "आपका भुगतान प्राप्त हो गया था।",
   "payments": "भुगतान",
   "pending": " (अपूर्ण)",
   "percentageOf": "${amount} का",
@@ -944,6 +944,13 @@
   "switchToEVMCompatibleWallet": "कृपया ईवीएम संगत वॉलेट पर स्विच करें और पुनः प्रयास करें (एथेरियम, पॉलीगॉन)",
   "symbol": "प्रतीक",
   "sync_all_wallets": "सभी वॉलेट सिंक करें",
+  "sync_attempting": "समन्वयन का प्रयास किया जा रहा है",
+  "sync_attempting_scan": "स्कैन का प्रयास किया जा रहा है",
+  "sync_connected": "जुड़े हुए",
+  "sync_connecting": "कनेक्ट",
+  "sync_not_connected": "जुड़े नहीं हैं",
+  "sync_offline": "ऑफलाइन",
+  "sync_starting_scan": "स्कैन प्रारंभ हो रहा है",
   "sync_status_attempting_scan": "स्कैन का प्रयास",
   "sync_status_attempting_sync": "सिंक करने का प्रयास",
   "sync_status_connected": "जुड़े हुए",
@@ -959,6 +966,9 @@
   "sync_status_syncronizing": "सिंक्रनाइज़ करने",
   "sync_status_timed_out": "समय समााप्त",
   "sync_status_unsupported": "असमर्थित नोड",
+  "sync_syncing": "सिंक्रनाइज़ किए जा रहे",
+  "sync_timed_out": "सिंक का समय समाप्त हो गया",
+  "sync_unsupported_node": "असमर्थित नोड",
   "synchronizing": "सिंक्रनाइज़ करना",
   "syncing_wallet_alert_content": "आपकी शेष राशि और लेनदेन सूची तब तक पूरी नहीं हो सकती जब तक कि शीर्ष पर \"सिंक्रनाइज़्ड\" न लिखा हो। अधिक जानने के लिए क्लिक/टैप करें।",
   "syncing_wallet_alert_title": "आपका वॉलेट सिंक हो रहा है",
