+++ res/values/strings_hr.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Prijeđite na novčanik kompatibilan s EVM-om i pokušajte ponovno (Ethereum, Polygon)",
   "symbol": "Simbol",
   "sync_all_wallets": "Sinkronizirajte sve novčanike",
+  "sync_attempting": "Pokušaj sinkronizacije",
+  "sync_attempting_scan": "Pokušaj skeniranja",
+  "sync_connected": "Povezan",
+  "sync_connecting": "Povezivanje",
+  "sync_not_connected": "Nije povezano",
+  "sync_offline": "Izvan mreže",
+  "sync_starting_scan": "Pokretanje skeniranja",
   "sync_status_attempting_scan": "Pokušaj skeniranja",
   "sync_status_attempting_sync": "POKUŠAJ SINKRONIZACIJE",
   "sync_status_connected": "SPOJENO",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "SINKRONIZIRANJE",
   "sync_status_timed_out": "ISTEKLO",
   "sync_status_unsupported": "Nepodržani čvor",
+  "sync_syncing": "Sinkronizacija",
+  "sync_timed_out": "Sinkronizacija je istekla",
+  "sync_unsupported_node": "Nepodržani čvor",
   "synchronizing": "Sinkronizirajući",
   "syncing_wallet_alert_content": "Vaš saldo i popis transakcija možda neće biti potpuni sve dok na vrhu ne piše \"SINKRONIZIRANO\". Kliknite/dodirnite da biste saznali više.",
   "syncing_wallet_alert_title": "Vaš novčanik se sinkronizira",
