+++ res/values/strings_hy.arb
@@ -940,6 +940,13 @@
   "switchToEVMCompatibleWallet": "Խնդրում ենք անցնել EVM համատեղելի դրամապանակ և փորձել կրկին (Ethereum, Polygon)",
   "symbol": "Նշան",
   "sync_all_wallets": "Համաժամեցնել բոլոր դրամապանակները",
+  "sync_attempting": "Համաժամացման փորձ",
+  "sync_attempting_scan": "Սկանավորման փորձ",
+  "sync_connected": "Միացված է",
+  "sync_connecting": "Միացում",
+  "sync_not_connected": "Միացված չէ",
+  "sync_offline": "Անցանց",
+  "sync_starting_scan": "Սկսվում է սկանավորումը",
   "sync_status_attempting_scan": "Փորձի սկան",
   "sync_status_attempting_sync": "ՀԱՄԱԺԱՄԵՑՄԱՆ ՓՈՐՁ",
   "sync_status_connected": "ՄԻԱՑՎԱԾԷ",
@@ -955,6 +962,9 @@
   "sync_status_syncronizing": "ՀԱՄԱԺԱՄԵՑՎՈՒՄ Է",
   "sync_status_timed_out": "Ժամանակը սպառվեց",
   "sync_status_unsupported": "ՉԱՋԱԿՑՎՈՂ ՀԱՆԳՈՒՅՑ,",
+  "sync_syncing": "Համաժամացում",
+  "sync_timed_out": "Համաժամացման ժամանակը սպառվել է",
+  "sync_unsupported_node": "Չաջակցվող հանգույց",
   "synchronizing": "Համաժամանակող",
   "syncing_wallet_alert_content": "Ձեր հաշիվը և գործարքների ցանկը կարող են լինել անավարտ մինչև վերին մասում գրված լինի “ՀԱՂՈՒՄ”։ Սեղմեք/դիպեք ուսումնասիրելու համար։",
   "syncing_wallet_alert_title": "Ձեր դրամապանակը համաժամացվում է",
