+++ res/values/strings_id.arb
@@ -945,6 +945,13 @@
   "switchToEVMCompatibleWallet": "Silakan beralih ke dompet yang kompatibel dengan EVM dan coba lagi (Ethereum, Polygon)",
   "symbol": "Simbol",
   "sync_all_wallets": "Sinkronkan semua dompet",
+  "sync_attempting": "Mencoba sinkronisasi",
+  "sync_attempting_scan": "Mencoba memindai",
+  "sync_connected": "Terhubung",
+  "sync_connecting": "Menghubungkan",
+  "sync_not_connected": "Tidak Terhubung",
+  "sync_offline": "Luring",
+  "sync_starting_scan": "Memulai pemindaian",
   "sync_status_attempting_scan": "Mencoba memindai",
   "sync_status_attempting_sync": "MENCOBA SINKRONISASI",
   "sync_status_connected": "TERHUBUNG",
@@ -960,6 +967,9 @@
   "sync_status_syncronizing": "SEDANG SINKRONISASI",
   "sync_status_timed_out": "WAKTU HABIS",
   "sync_status_unsupported": "Node yang tidak didukung",
+  "sync_syncing": "Sinkronisasi",
+  "sync_timed_out": "Waktu Sinkronisasi Habis",
+  "sync_unsupported_node": "Node tidak didukung",
   "synchronizing": "Sinkronisasi",
   "syncing_wallet_alert_content": "Saldo dan daftar transaksi Anda mungkin belum lengkap sampai tertulis “SYNCHRONIZED” di bagian atas. Klik/ketuk untuk mempelajari lebih lanjut.",
   "syncing_wallet_alert_title": "Dompet Anda sedang disinkronkan",
