+++ res/values/strings_it.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Passa a un portafoglio compatibile con EVM e riprova (Ethereum, Polygon)",
   "symbol": "Simbolo",
   "sync_all_wallets": "Sincronizza tutti i portafogli",
+  "sync_attempting": "Tentativo di sincronizzazione",
+  "sync_attempting_scan": "Tentativo di scansione",
+  "sync_connected": "Collegato",
+  "sync_connecting": "Connessione",
+  "sync_not_connected": "Non connesso",
+  "sync_offline": "Non in linea",
+  "sync_starting_scan": "Avvio della scansione",
   "sync_status_attempting_scan": "TENTATIVO DI SCANSIONE",
   "sync_status_attempting_sync": "TENTATIVO DI SINCRONIZZAZIONE",
   "sync_status_connected": "CONNESSO",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "SINCRONIZZAZIONE",
   "sync_status_timed_out": "TIMED OUT",
   "sync_status_unsupported": "NODO NON SUPPORTATO",
+  "sync_syncing": "Sincronizzazione",
+  "sync_timed_out": "Sincronizzazione scaduta",
+  "sync_unsupported_node": "Nodo non supportato",
   "synchronizing": "Sincronizzazione",
   "syncing_wallet_alert_content": "Il saldo e l'elenco delle transazioni potrebbero non essere completi fin quando non viene visualizzato \"SINCRONIZZATO\" in alto. Clicca/tocca per saperne di più.",
   "syncing_wallet_alert_title": "Il tuo portafoglio si sta sincronizzando",
