+++ res/values/strings_ja.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "EVM 互換のウォレットに切り替えて再試行してください (イーサリアム、ポリゴン)",
   "symbol": "シンボル",
   "sync_all_wallets": "すべてのウォレットを同期",
+  "sync_attempting": "同期を試みています",
+  "sync_attempting_scan": "スキャンを試みています",
+  "sync_connected": "接続済み",
+  "sync_connecting": "接続中",
+  "sync_not_connected": "接続されていません",
+  "sync_offline": "オフライン",
+  "sync_starting_scan": "スキャンの開始",
   "sync_status_attempting_scan": "スキャンの試み",
   "sync_status_attempting_sync": "同期を試みています",
   "sync_status_connected": "接続済み",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "同期",
   "sync_status_timed_out": "タイムアウトしました",
   "sync_status_unsupported": "サポートされていないノード",
+  "sync_syncing": "同期中",
+  "sync_timed_out": "同期がタイムアウトしました",
+  "sync_unsupported_node": "サポートされていないノード",
   "synchronizing": "同期",
   "syncing_wallet_alert_content": "上部に「同期済み」と表示されるまで、残高と取引リストが完了していない可能性があります。詳細については、クリック/タップしてください。",
   "syncing_wallet_alert_title": "ウォレットは同期中です",
