+++ res/values/strings_ko.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "EVM 호환 지갑(이더리움, 폴리곤)으로 전환하고 다시 시도하세요",
   "symbol": "심볼",
   "sync_all_wallets": "모든 지갑 동기화",
+  "sync_attempting": "동기화 시도 중",
+  "sync_attempting_scan": "스캔을 시도하는 중",
+  "sync_connected": "연결됨",
+  "sync_connecting": "연결 중",
+  "sync_not_connected": "연결되지 않음",
+  "sync_offline": "오프라인",
+  "sync_starting_scan": "스캔 시작",
   "sync_status_attempting_scan": "스캔 시도 중",
   "sync_status_attempting_sync": "동기화 시도 중",
   "sync_status_connected": "연결됨",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "동기화 중",
   "sync_status_timed_out": "시간 초과",
   "sync_status_unsupported": "지원되지 않는 노드",
+  "sync_syncing": "동기화 중",
+  "sync_timed_out": "동기화 시간이 초과되었습니다",
+  "sync_unsupported_node": "지원되지 않는 노드",
   "synchronizing": "동기화 중",
   "syncing_wallet_alert_content": "상단에 “동기화됨”이라고 표시될 때까지 잔액 및 트랜잭션 목록이 완전하지 않을 수 있습니다. 자세히 알아보려면 클릭/탭하세요.",
   "syncing_wallet_alert_title": "지갑 동기화 중",
