+++ res/values/strings_my.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "ကျေးဇူးပြု၍ EVM တွဲဖက်သုံးနိုင်သော ပိုက်ဆံအိတ်သို့ ပြောင်းပြီး ထပ်စမ်းကြည့်ပါ (Ethereum၊ Polygon)",
   "symbol": "သင်္ကေတ",
   "sync_all_wallets": "အားလုံးပိုက်ဆံအိတ်စည်းညှိ",
+  "sync_attempting": "ထပ်တူကျအောင် ကြိုးစားနေသည်။",
+  "sync_attempting_scan": "စကင်န်ကြိုးစားခြင်း။",
+  "sync_connected": "ချိတ်ဆက်ထားသည်။",
+  "sync_connecting": "ချိတ်ဆက်နေသည်။",
+  "sync_not_connected": "မချိတ်ဆက်ပါ။",
+  "sync_offline": "အော့ဖ်လိုင်း",
+  "sync_starting_scan": "စကင်န်ဖတ်ခြင်း",
   "sync_status_attempting_scan": "scan ကြိုးစားနေ",
   "sync_status_attempting_sync": "ချိန်ကိုက်ခြင်းကို ကြိုးစားနေသည်။",
   "sync_status_connected": "ချိတ်ဆက်ထားသည်။",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "ထပ်တူပြုခြင်း။",
   "sync_status_timed_out": "ထွက်အချိန်ကုန်",
   "sync_status_unsupported": "node မထောက်ပံ့ node ကို",
+  "sync_syncing": "ထပ်တူပြုခြင်း။",
+  "sync_timed_out": "စင့်ခ်လုပ်ချိန်ကုန်သွားပါပြီ။",
+  "sync_unsupported_node": "ပံ့ပိုးမထားသော node",
   "synchronizing": "ထပ်တူပြုခြင်း",
   "syncing_wallet_alert_content": "သင်၏လက်ကျန်နှင့် ငွေပေးငွေယူစာရင်းသည် ထိပ်တွင် \"Synchronizeed\" ဟုပြောသည်အထိ မပြီးမြောက်နိုင်ပါ။ ပိုမိုလေ့လာရန် နှိပ်/နှိပ်ပါ။",
   "syncing_wallet_alert_title": "သင့်ပိုက်ဆံအိတ်ကို စင့်ခ်လုပ်နေပါသည်။",
