+++ res/values/strings_nl.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Schakel over naar een EVM-compatibele portemonnee en probeer het opnieuw (Ethereum, Polygon)",
   "symbol": "Symbool",
   "sync_all_wallets": "Alle portemonnees synchroniseren",
+  "sync_attempting": "Er wordt geprobeerd te synchroniseren",
+  "sync_attempting_scan": "Proberen te scannen",
+  "sync_connected": "Aangesloten",
+  "sync_connecting": "Verbinden",
+  "sync_not_connected": "Niet verbonden",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Scan starten",
   "sync_status_attempting_scan": "Proberen scan",
   "sync_status_attempting_sync": "SYNCHRONISATIE PROBEREN",
   "sync_status_connected": "VERBONDEN",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "SYNCHRONISEREN",
   "sync_status_timed_out": "Uitgeput",
   "sync_status_unsupported": "Niet ondersteund knooppunt",
+  "sync_syncing": "Synchroniseren",
+  "sync_timed_out": "Er is een time-out voor de synchronisatie opgetreden",
+  "sync_unsupported_node": "Niet-ondersteund knooppunt",
   "synchronizing": "Synchronisatie",
   "syncing_wallet_alert_content": "Uw saldo- en transactielijst is mogelijk pas compleet als er bovenaan 'GESYNCHRONISEERD' staat. Klik/tik voor meer informatie.",
   "syncing_wallet_alert_title": "Uw portemonnee wordt gesynchroniseerd",
