+++ res/values/strings_pl.arb
@@ -941,6 +941,13 @@
   "switchToEVMCompatibleWallet": "Przejdź na portfel kompatybilny z EVM i spróbuj ponownie (Ethereum, Polygon)",
   "symbol": "Symbol",
   "sync_all_wallets": "Synchronizuj wszystkie portfele",
+  "sync_attempting": "Próba synchronizacji",
+  "sync_attempting_scan": "Próba skanowania",
+  "sync_connected": "Połączony",
+  "sync_connecting": "Złączony",
+  "sync_not_connected": "Brak połączenia",
+  "sync_offline": "Nieaktywny",
+  "sync_starting_scan": "Rozpoczęcie skanowania",
   "sync_status_attempting_scan": "Rozpoczynanie skanowania",
   "sync_status_attempting_sync": "PRÓBA SYCHRONIZACJI",
   "sync_status_connected": "POŁĄCZONO",
@@ -956,6 +963,9 @@
   "sync_status_syncronizing": "SYNCHRONIZOWANIE",
   "sync_status_timed_out": "PRZEKROCZONO LIMIT CZASU",
   "sync_status_unsupported": "Nieobsługiwany węzeł",
+  "sync_syncing": "Synchronizowanie",
+  "sync_timed_out": "Upłynął limit czasu synchronizacji",
+  "sync_unsupported_node": "Nieobsługiwany węzeł",
   "synchronizing": "Synchronizacja",
   "syncing_wallet_alert_content": "Twoje saldo i lista transakcji mogą nie być kompletne, dopóki nie pojawi się na górze napis „SYNCHRONIZED”. Kliknij/stuknij, aby dowiedzieć się więcej.",
   "syncing_wallet_alert_title": "Twój portfel się synchronizuje",
