+++ res/values/strings_pt.arb
@@ -944,6 +944,13 @@
   "switchToEVMCompatibleWallet": "Mude para uma carteira compatível com EVM e tente novamente (Ethereum, Polygon)",
   "symbol": "Símbolo",
   "sync_all_wallets": "Sincronize todas as carteiras",
+  "sync_attempting": "Tentando sincronizar",
+  "sync_attempting_scan": "Tentando verificar",
+  "sync_connected": "Conectado",
+  "sync_connecting": "Conectando",
+  "sync_not_connected": "Não conectado",
+  "sync_offline": "Off-line",
+  "sync_starting_scan": "Iniciando verificação",
   "sync_status_attempting_scan": "TENTANDO ESCANEAR",
   "sync_status_attempting_sync": "TENTANDO SINCRONIZAR",
   "sync_status_connected": "CONECTADO",
@@ -959,6 +966,9 @@
   "sync_status_syncronizing": "SINCRONIZANDO",
   "sync_status_timed_out": "TEMPO ESGOTADO",
   "sync_status_unsupported": "Nó não suportado",
+  "sync_syncing": "Sincronizando",
+  "sync_timed_out": "A sincronização expirou",
+  "sync_unsupported_node": "Nó não compatível",
   "synchronizing": "Sincronizando",
   "syncing_wallet_alert_content": "Seu saldo e lista de transações podem não estar completos até que diga “SYNCHRONIZED” no topo. Clique/toque para saber mais.",
   "syncing_wallet_alert_title": "Sua carteira está sincronizando",
