+++ res/values/strings_ru.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Пожалуйста, переключитесь на кошелек, совместимый с EVM, и повторите попытку (Ethereum, Polygon).",
   "symbol": "Символ",
   "sync_all_wallets": "Синхронизировать все кошельки",
+  "sync_attempting": "Попытка синхронизации",
+  "sync_attempting_scan": "Попытка сканирования",
+  "sync_connected": "Подключено",
+  "sync_connecting": "Подключение",
+  "sync_not_connected": "Не подключен",
+  "sync_offline": "Оффлайн",
+  "sync_starting_scan": "Запуск сканирования",
   "sync_status_attempting_scan": "Попытка сканирования",
   "sync_status_attempting_sync": "ПОПЫТКА СИНХРОНИЗАЦИИ",
   "sync_status_connected": "ПОДКЛЮЧЕНО",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "СИНХРОНИЗАЦИЯ",
   "sync_status_timed_out": "ВРЕМЯ ВЫШЛО",
   "sync_status_unsupported": "Неподдерживаемый узел",
+  "sync_syncing": "Синхронизация",
+  "sync_timed_out": "Время синхронизации истекло",
+  "sync_unsupported_node": "Неподдерживаемый узел",
   "synchronizing": "Синхронизация",
   "syncing_wallet_alert_content": "Ваш баланс и список транзакций могут быть неполными, пока вверху не будет написано «СИНХРОНИЗИРОВАНО». Щелкните/коснитесь, чтобы узнать больше.",
   "syncing_wallet_alert_title": "Ваш кошелек синхронизируется",
