+++ res/values/strings_th.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "โปรดเปลี่ยนไปใช้กระเป๋าเงินที่รองรับ EVM แล้วลองอีกครั้ง (Ethereum, Polygon)",
   "symbol": "เครื่องหมาย",
   "sync_all_wallets": "ซิงค์กระเป๋าเงินทั้งหมด",
+  "sync_attempting": "กำลังพยายามซิงค์",
+  "sync_attempting_scan": "กำลังพยายามสแกน",
+  "sync_connected": "เชื่อมต่อแล้ว",
+  "sync_connecting": "กำลังเชื่อมต่อ",
+  "sync_not_connected": "ไม่ได้เชื่อมต่อ",
+  "sync_offline": "ออฟไลน์",
+  "sync_starting_scan": "กำลังเริ่มการสแกน",
   "sync_status_attempting_scan": "พยายามสแกน",
   "sync_status_attempting_sync": "พยายามซิงโครไนซ์",
   "sync_status_connected": "เชื่อมต่อแล้ว",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "กำลังซิงโครไนซ์",
   "sync_status_timed_out": "หมดเวลา",
   "sync_status_unsupported": "โหนดที่ไม่ได้รับการสนับสนุน",
+  "sync_syncing": "กำลังซิงค์",
+  "sync_timed_out": "การซิงค์หมดเวลา",
+  "sync_unsupported_node": "โหนดที่ไม่รองรับ",
   "synchronizing": "การซิงโครไนซ์",
   "syncing_wallet_alert_content": "รายการยอดเงินและธุรกรรมของคุณอาจไม่สมบูรณ์จนกว่าจะมีข้อความว่า “ซิงโครไนซ์” ที่ด้านบน คลิก/แตะเพื่อเรียนรู้เพิ่มเติม่",
   "syncing_wallet_alert_title": "กระเป๋าสตางค์ของคุณกำลังซิงค์",
