+++ res/values/strings_tl.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Mangyaring lumipat sa isang EVM compatible na wallet at subukang muli (Ethereum, Polygon)",
   "symbol": "Simbolo",
   "sync_all_wallets": "I-sync ang lahat ng mga wallet",
+  "sync_attempting": "Sinusubukang i-sync",
+  "sync_attempting_scan": "Sinusubukang i-scan",
+  "sync_connected": "Nakakonekta",
+  "sync_connecting": "Kumokonekta",
+  "sync_not_connected": "Hindi Konektado",
+  "sync_offline": "Offline",
+  "sync_starting_scan": "Nagsisimula sa pag-scan",
   "sync_status_attempting_scan": "Pagtatangka ng pag -scan",
   "sync_status_attempting_sync": "SINUSUBUKANG I-SYNC",
   "sync_status_connected": "KONEKTADO",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "PAG-SYNCHRONIZE",
   "sync_status_timed_out": "NAG-TIME OUT",
   "sync_status_unsupported": "HINDI SUPORTADONG NODE",
+  "sync_syncing": "Nagsi-sync",
+  "sync_timed_out": "Nag-time Out ang Sync",
+  "sync_unsupported_node": "Hindi sinusuportahang node",
   "synchronizing": "Pag -synchronize",
   "syncing_wallet_alert_content": "Ang iyong balanse at listahan ng transaksyon ay maaaring hindi kumpleto hanggang sa sabihin nito na \"NAKA-SYNCHRONIZE\" sa tuktok. Mag-click/tap upang malaman ang higit pa.",
   "syncing_wallet_alert_title": "Ang iyong wallet ay nag-sync",
