+++ res/values/strings_tr.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "Lütfen EVM uyumlu bir cüzdana geçin ve tekrar deneyin (Ethereum, Polygon)",
   "symbol": "Sembol",
   "sync_all_wallets": "Tüm cüzdanları senkronize edin",
+  "sync_attempting": "Senkronizasyon deneniyor",
+  "sync_attempting_scan": "Tarama deneniyor",
+  "sync_connected": "Bağlı",
+  "sync_connecting": "Bağlanıyor",
+  "sync_not_connected": "Bağlı Değil",
+  "sync_offline": "Çevrimdışı",
+  "sync_starting_scan": "Tarama başlatılıyor",
   "sync_status_attempting_scan": "Tarama deneme",
   "sync_status_attempting_sync": "SENKRONİZE EDİLMEYE ÇALIŞILIYOR",
   "sync_status_connected": "BAĞLANILDI",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "SENKRONİZE EDİLİYOR",
   "sync_status_timed_out": "ZAMAN AŞIMINA UĞRADI",
   "sync_status_unsupported": "Desteklenmeyen düğüm",
+  "sync_syncing": "Senkronizasyon",
+  "sync_timed_out": "Senkronizasyon Zaman Aşımına Uğradı",
+  "sync_unsupported_node": "Desteklenmeyen düğüm",
   "synchronizing": "Senkronize etme",
   "syncing_wallet_alert_content": "Bakiyeniz ve işlem listeniz, en üstte \"SENKRONİZE EDİLDİ\" yazana kadar tamamlanmamış olabilir. Daha fazla bilgi edinmek için tıklayın/dokunun.",
   "syncing_wallet_alert_title": "Cüzdanınız senkronize ediliyor",
