+++ res/values/strings_uk.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Перейдіть на гаманець, сумісний з EVM, і повторіть спробу (Ethereum, Polygon)",
   "symbol": "символ",
   "sync_all_wallets": "Синхронізувати всі гаманці",
+  "sync_attempting": "Спроба синхронізації",
+  "sync_attempting_scan": "Спроба сканування",
+  "sync_connected": "Підключено",
+  "sync_connecting": "Підключення",
+  "sync_not_connected": "Не підключено",
+  "sync_offline": "Офлайн",
+  "sync_starting_scan": "Початок сканування",
   "sync_status_attempting_scan": "Спроба сканування",
   "sync_status_attempting_sync": "СПРОБА СИНХРОНІЗАЦІЇ",
   "sync_status_connected": "ПІДКЛЮЧЕНО",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "СИНХРОНІЗАЦІЯ",
   "sync_status_timed_out": "ТАЙМ-АУТ",
   "sync_status_unsupported": "Непідтримуваний вузол",
+  "sync_syncing": "Синхронізація",
+  "sync_timed_out": "Час очікування синхронізації минув",
+  "sync_unsupported_node": "Непідтримуваний вузол",
   "synchronizing": "Синхронізація",
   "syncing_wallet_alert_content": "Ваш баланс та список транзакцій може бути неповним, доки вгорі не буде написано «СИНХРОНІЗОВАНО». Натисніть/торкніться, щоб дізнатися більше.",
   "syncing_wallet_alert_title": "Ваш гаманець синхронізується",
