+++ res/values/strings_ur.arb
@@ -944,6 +944,13 @@
   "switchToEVMCompatibleWallet": "(Ethereum, Polygon) ﮟﯾﺮﮐ ﺶﺷﻮﮐ ﮦﺭﺎﺑﻭﺩ ﺭﻭﺍ ﮟﯾﺮﮐ ﭻﺋﻮﺳ ﺮﭘ ﭧﯿﻟﺍﻭ ﮯﻟﺍﻭ ﮯﻨﮭﮐﺭ ﺖﻘﺑﺎﻄﻣ ",
   "symbol": "ﺖﻣﻼ﻿ﻋ",
   "sync_all_wallets": "تمام بٹوے کو ہم آہنگ کریں",
+  "sync_attempting": "مطابقت پذیری کی کوشش کرنا",
+  "sync_attempting_scan": "اسکین کی کوشش کرنا",
+  "sync_connected": "منسلک",
+  "sync_connecting": "رابطہ قائم کرنا",
+  "sync_not_connected": "منسلک نہیں",
+  "sync_offline": "آف لائن",
+  "sync_starting_scan": "اسکین شروع کرنا",
   "sync_status_attempting_scan": "اسکین کی کوشش کرنا",
   "sync_status_attempting_sync": "ہم آہنگی کی کوشش کر رہا ہے۔",
   "sync_status_connected": "منسلک",
@@ -959,6 +966,9 @@
   "sync_status_syncronizing": "مطابقت پذیری",
   "sync_status_timed_out": "وقت ختم",
   "sync_status_unsupported": "غیر تعاون یافتہ نوڈ",
+  "sync_syncing": "مطابقت پذیری",
+  "sync_timed_out": "مطابقت پذیری کا وقت ختم ہوگیا",
+  "sync_unsupported_node": "غیر تعاون شدہ نوڈ",
   "synchronizing": "ہم آہنگی",
   "syncing_wallet_alert_content": "آپ کے بیلنس اور لین دین کی فہرست اس وقت تک مکمل نہیں ہو سکتی جب تک کہ یہ سب سے اوپر \"SYNCRONIZED\" نہ کہے۔ مزید جاننے کے لیے کلک/تھپتھپائیں۔",
   "syncing_wallet_alert_title": "آپ کا بٹوہ مطابقت پذیر ہو رہا ہے۔",
