+++ res/values/strings_vi.arb
@@ -939,6 +939,13 @@
   "switchToEVMCompatibleWallet": "Vui lòng chuyển sang ví tương thích EVM và thử lại (Ethereum, Polygon)",
   "symbol": "Ký hiệu",
   "sync_all_wallets": "Đồng bộ tất cả các ví",
+  "sync_attempting": "Đang cố gắng đồng bộ hóa",
+  "sync_attempting_scan": "Đang cố gắng quét",
+  "sync_connected": "Đã kết nối",
+  "sync_connecting": "Đang kết nối",
+  "sync_not_connected": "Chưa kết nối",
+  "sync_offline": "Ngoại tuyến",
+  "sync_starting_scan": "Bắt đầu quét",
   "sync_status_attempting_scan": "Cố gắng quét",
   "sync_status_attempting_sync": "ĐANG THỬ ĐỒNG BỘ",
   "sync_status_connected": "ĐÃ KẾT NỐI",
@@ -954,6 +961,9 @@
   "sync_status_syncronizing": "ĐANG ĐỒNG BỘ",
   "sync_status_timed_out": "HẾT THỜI GIAN",
   "sync_status_unsupported": "NÓT KHÔNG ĐƯỢC HỖ TRỢ",
+  "sync_syncing": "Đang đồng bộ hóa",
+  "sync_timed_out": "Đã hết thời gian đồng bộ hóa",
+  "sync_unsupported_node": "Nút không được hỗ trợ",
   "synchronizing": "Đồng bộ hóa",
   "syncing_wallet_alert_content": "Số dư và danh sách giao dịch của bạn có thể không đầy đủ cho đến khi nó hiển thị “ĐÃ ĐỒNG BỘ” ở trên cùng. Nhấn vào đây để tìm hiểu thêm.",
   "syncing_wallet_alert_title": "Ví của bạn đang đồng bộ",
