+++ res/values/strings_yo.arb
@@ -943,6 +943,13 @@
   "switchToEVMCompatibleWallet": "Jọwọ yipada si apamọwọ ibaramu EVM ki o tun gbiyanju lẹẹkansi (Ethereum, Polygon)",
   "symbol": "Aami",
   "sync_all_wallets": "Muṣiṣẹpọ gbogbo awọn Woleti",
+  "sync_attempting": "Igbiyanju amuṣiṣẹpọ",
+  "sync_attempting_scan": "Igbiyanju ọlọjẹ",
+  "sync_connected": "Ti sopọ",
+  "sync_connecting": "Nsopọ",
+  "sync_not_connected": "Ko Sopọ",
+  "sync_offline": "Aisinipo",
+  "sync_starting_scan": "Ibẹrẹ ọlọjẹ",
   "sync_status_attempting_scan": "Igbiyanju ọlọjẹ",
   "sync_status_attempting_sync": "Ń GBÌYÀNJÚ MÚDỌ́GBA",
   "sync_status_connected": "TI DÁRAPỌ̀ MỌ́",
@@ -958,6 +965,9 @@
   "sync_status_syncronizing": "Ń MÚDỌ́GBA",
   "sync_status_timed_out": "Ti akoko jade",
   "sync_status_unsupported": "Ile-igbimọ ti ko ni atilẹyin",
+  "sync_syncing": "Mimuuṣiṣẹpọ",
+  "sync_timed_out": "Aago amuṣiṣẹpọ",
+  "sync_unsupported_node": "Node ti ko ni atilẹyin",
   "synchronizing": "Mimuuṣiṣẹpọ",
   "syncing_wallet_alert_content": "Iwontunws.funfun rẹ ati atokọ idunadura le ma pari titi ti yoo fi sọ “SYNCHRONIZED” ni oke. Tẹ/tẹ ni kia kia lati ni imọ siwaju sii.",
   "syncing_wallet_alert_title": "Apamọwọ rẹ n muṣiṣẹpọ",
