+++ res/values/strings_zh.arb
@@ -942,6 +942,13 @@
   "switchToEVMCompatibleWallet": "请切换到 EVM 兼容钱包并重试（以太坊、Polygon）",
   "symbol": "符号",
   "sync_all_wallets": "同步所有钱包",
+  "sync_attempting": "尝试同步",
+  "sync_attempting_scan": "正在尝试扫描",
+  "sync_connected": "已连接",
+  "sync_connecting": "连接中",
+  "sync_not_connected": "未连接",
+  "sync_offline": "离线",
+  "sync_starting_scan": "开始扫描",
   "sync_status_attempting_scan": "正在尝试扫描",
   "sync_status_attempting_sync": "正在尝试同步",
   "sync_status_connected": "已连接",
@@ -957,6 +964,9 @@
   "sync_status_syncronizing": "正在同步",
   "sync_status_timed_out": "超时",
   "sync_status_unsupported": "不支持的节点",
+  "sync_syncing": "正在同步",
+  "sync_timed_out": "同步超时",
+  "sync_unsupported_node": "不支持的节点",
   "synchronizing": "同步",
   "syncing_wallet_alert_content": "同步中，余额和交易列表可能不完整，直至顶部显示“已同步”。点击了解详情。",
   "syncing_wallet_alert_title": "您的钱包正在同步",
